UBI: simplify the find_consolidable_lebs() logic
We don't have remove the LEB from the full list when searching for
consolidable LEBs, because we already hold the lock on the LEB, thus
preventing any change on the full LEB entry while we are consolidating.

Full LEB entries will be invalidated by consolidate_lebs() at the end of
the consolidation process, and will remain in the full list if the process
is interrupted.

Signed-off-by: Boris Brezillon <boris.brezillon@free-electrons.com>